refactor(dev-preview): extract diagnostics ring and disk-usage helpers
- Move the per-session diagnostics ring buffer (constants, capDiagnosticText,
  recordDiagnostic) into DevPreviewDiagnosticsRing.ts, mirroring the
  Map-parameter style already used by DevPreviewPortAllocator.ts
- Move disk-usage and package-manager detection (CACHE_DIRS, clearCacheDirs,
  detectPackageManagerInfo, detectInstallCommand, statDirMeta, computeDirSize)
  into DevPreviewDiskUsage.ts as pure functions
- DevPreviewSessionService delegates to both via thin wrappers; public API
  and DIAGNOSTIC_RING_MAX re-export unchanged
