+++ electron/services/DevPreviewSessionService.ts
@@ -1,5 +1,3 @@
-import { execFile } from "node:child_process";
-import { existsSync } from "node:fs";
 import fsPromises from "node:fs/promises";
 import path from "node:path";
 import { stripVTControlCharacters } from "node:util";
@@ -8,6 +6,20 @@ import {
   normalizeNextjsDevCommand,
   normalizeViteDevCommand,
 } from "./DevPreviewCommandNormalizer.js";
+import {
+  CACHE_DIRS,
+  clearCacheDirs,
+  computeDirSize,
+  detectInstallCommand,
+  detectPackageManagerInfo,
+  statDirMeta,
+} from "./DevPreviewDiskUsage.js";
+import {
+  capDiagnosticText,
+  recordDevPreviewDiagnostic,
+  type DevPreviewDiagnosticInput,
+  type DevPreviewDiagnosticsRingMap,
+} from "./DevPreviewDiagnosticsRing.js";
 import {
   allocatePort,
   releasePort,
@@ -26,6 +38,7 @@ import {
 } from "./DevPreviewRequestValidators.js";
 
 export { normalizeNextjsDevCommand } from "./DevPreviewCommandNormalizer.js";
+export { DIAGNOSTIC_RING_MAX } from "./DevPreviewDiagnosticsRing.js";
 import { buildDevPreviewSubdomain } from "../../shared/utils/devPreviewProxy.js";
 import type { DevPreviewManifestEntry } from "./DevPreviewManifestService.js";
 import type { PtyClient } from "./PtyClient.js";
@@ -40,11 +53,9 @@ import type {
   DevPreviewDiagnosticsSnapshot,
   DevPreviewProxyFailureCause,
   DevPreviewUpstreamResolution,
-  DevPreviewDirMeta,
   DevPreviewDestructivePreviewMeta,
   DevPreviewDestructivePreviewSizes,
   DevPreviewDestructivePreviewSizesRequest,
-  DevPreviewPackageManager,
 } from "../../shared/types/ipc/devPreview.js";
 import type { DevServerError } from "../../shared/utils/devServerErrors.js";
 import { detectDevServerError } from "../../shared/utils/devServerErrors.js";
@@ -91,20 +102,6 @@ const RUNNING_STATES: ReadonlySet<DevPreviewSessionStatus> = new Set([
   "running",
 ]);
 
-// Framework build/dep caches wiped by restartAndClearCache, relative to the
-// session cwd. Covers Next, Vite, Turbo, SvelteKit, Astro, and Nuxt root caches
-// plus Vite's node_modules/.vite dep-optimization cache (distinct from a
-// root-level .vite directory). Missing dirs are tolerated by removal callers.
-const CACHE_DIRS: readonly string[] = [
-  ".next",
-  ".vite",
-  ".turbo",
-  ".svelte-kit",
-  ".astro",
-  ".nuxt",
-  path.join("node_modules", ".vite"),
-];
-
 const DEFAULT_TIMEOUT_MS = 8000;
 const DEV_PREVIEW_STOP_ESCALATION_MS = 5000;
 const STALE_START_RECOVERY_MS = 10000;
@@ -135,30 +132,6 @@ const CRASH_LOOP_INITIAL_DELAY_MS = 500;
 const CRASH_LOOP_BACKOFF_MULTIPLIER = 1.5;
 const CRASH_LOOP_MAX_DELAY_MS = 3000;
 const CRASH_LOOP_MAX_ATTEMPTS = 5;
-
-// Bounds for the per-session diagnostics timeline. The ring keeps the last
-// DIAGNOSTIC_RING_MAX events per session key; the map keeps rings for the last
-// DIAGNOSTIC_SESSIONS_MAX keys touched so a timeline survives hibernation /
-// panel-close (the session object is deleted but its ring stays inspectable)
-// without growing unbounded across a long app session. Free-text fields are
-// capped so one pathological error line can't bloat a snapshot.
-export const DIAGNOSTIC_RING_MAX = 100;
-const DIAGNOSTIC_SESSIONS_MAX = 50;
-const DIAGNOSTIC_TEXT_MAX = 200;
-// Consecutive identical proxy failures coalesce into one event with a count —
-// a webview retry storm must not evict the lifecycle history it explains.
-const PROXY_EVENT_COALESCE_MS = 10_000;
-
-// A diagnostic event as recorded at a call site: the ring assigns at/seq/
-// generation/count. Distributes over the union so each variant keeps its
-// typed details.
-type DiagnosticInput<T = DevPreviewDiagnosticEvent> = T extends DevPreviewDiagnosticEvent
-  ? Omit<T, "at" | "seq" | "generation" | "count">
-  : never;
-
-function capDiagnosticText(text: string): string {
-  return text.length > DIAGNOSTIC_TEXT_MAX ? text.slice(0, DIAGNOSTIC_TEXT_MAX) : text;
-}
 
 /**
  * Parse both the port AND the transport scheme out of a detected dev-server URL (#9974).
@@ -202,10 +175,7 @@ export class DevPreviewSessionService {
   // separate map (not on the session object) so the timeline outlives session
   // deletion — a "why did my server stop?" after hibernation or panel-close is
   // exactly what this exists to answer. LRU-bounded via re-insertion on write.
-  private readonly diagnostics = new Map<
-    string,
-    { events: DevPreviewDiagnosticEvent[]; seq: number }
-  >();
+  private readonly diagnostics: DevPreviewDiagnosticsRingMap = new Map();
   // In-flight waitForPortFree aborts, so dispose() doesn't leave locked tasks
   // polling a busy port for up to PORT_FREE_TIMEOUT_MS after app quit began.
   private readonly portWaitAborts = new Set<AbortController>();
@@ -620,7 +590,7 @@ export class DevPreviewSessionService {
         return;
       }
 
-      const deletionError = await this.clearCacheDirs(session.cwd);
+      const deletionError = await clearCacheDirs(session.cwd);
       if (deletionError) {
         this.updateSession(session, {
           status: "error",
@@ -707,23 +677,6 @@ export class DevPreviewSessionService {
       await this.runInstall(session);
     });
     return this.getSessionState(request.projectId, request.panelId);
-  }
-
-  /**
-   * Deletes the framework cache directories under cwd. Returns an error
-   * message string if any deletion failed, or null on success. `force: true`
-   * makes missing directories a no-op.
-   */
-  private async clearCacheDirs(cwd: string): Promise<string | null> {
-    const failures: string[] = [];
-    for (const dir of CACHE_DIRS) {
-      try {
-        await fsPromises.rm(path.join(cwd, dir), { recursive: true, force: true });
-      } catch (err) {
-        failures.push(`${dir} (${formatErrorMessage(err, "deletion failed")})`);
-      }
-    }
-    return failures.length > 0 ? failures.join(", ") : null;
   }
 
   async stop(request: DevPreviewSessionRequest): Promise<DevPreviewSessionState> {
@@ -1048,55 +1001,22 @@ export class DevPreviewSessionService {
   }
 
   /**
-   * Append a diagnostic event to a session key's bounded ring. Consecutive
-   * identical proxy failures coalesce into one event with a count so a webview
-   * retry storm can't evict the lifecycle history. Writing touches the key's
-   * LRU position; the oldest untouched key is dropped past DIAGNOSTIC_SESSIONS_MAX.
+   * Append a diagnostic event to a session key's bounded ring. See
+   * `recordDevPreviewDiagnostic` for coalescing/eviction behavior.
    */
-  private recordDiagnostic(key: string, generation: number, input: DiagnosticInput): void {
+  private recordDiagnostic(
+    key: string,
+    generation: number,
+    input: DevPreviewDiagnosticInput
+  ): void {
     if (this.disposed) return;
-    let ring = this.diagnostics.get(key);
-    if (ring) {
-      this.diagnostics.delete(key);
-    } else {
-      ring = { events: [], seq: 0 };
-    }
-    this.diagnostics.set(key, ring);
-    while (this.diagnostics.size > DIAGNOSTIC_SESSIONS_MAX) {
-      const oldest = this.diagnostics.keys().next().value;
-      if (oldest === undefined) break;
-      this.diagnostics.delete(oldest);
-    }
-
-    // Coalesce only true repeats: same failure, same errno, same generation —
-    // a post-restart failure or a different errno must stay its own event.
-    const last = ring.events[ring.events.length - 1];
-    if (
-      (input.type === "proxy-502" || input.type === "proxy-ws-failed") &&
-      last !== undefined &&
-      last.type === input.type &&
-      last.cause === input.cause &&
-      last.code === input.code &&
-      last.generation === generation &&
-      Date.now() - last.at <= PROXY_EVENT_COALESCE_MS
-    ) {
-      last.count = (last.count ?? 1) + 1;
-      last.at = Date.now();
-      return;
-    }
-
-    ring.events.push({
-      ...input,
-      at: Date.now(),
-      seq: ring.seq++,
-      generation,
-    } as DevPreviewDiagnosticEvent);
-    if (ring.events.length > DIAGNOSTIC_RING_MAX) {
-      ring.events.splice(0, ring.events.length - DIAGNOSTIC_RING_MAX);
-    }
+    recordDevPreviewDiagnostic(this.diagnostics, key, generation, input);
   }
 
-  private recordSessionDiagnostic(session: DevPreviewSession, input: DiagnosticInput): void {
+  private recordSessionDiagnostic(
+    session: DevPreviewSession,
+    input: DevPreviewDiagnosticInput
+  ): void {
     this.recordDiagnostic(
       createSessionKey(session.projectId, session.panelId),
       session.generation,
@@ -2143,7 +2063,7 @@ export class DevPreviewSessionService {
     session.installAttemptedGeneration = session.generation;
     session.isRunningInstall = true;
 
-    const installCommand = this.detectInstallCommand(session.cwd);
+    const installCommand = detectInstallCommand(session.cwd);
     this.recordSessionDiagnostic(session, {
       type: "install-started",
       command: capDiagnosticText(installCommand),
@@ -2203,47 +2123,16 @@ export class DevPreviewSessionService {
     }, PTY_SUBMIT_AFTER_SPAWN_MS);
   }
 
-  private detectInstallCommand(cwd: string): string {
-    const info = this.detectPackageManagerInfo(cwd);
-    return `${info.packageManager} install`;
-  }
-
-  // Lockfile precedence mirrors detectInstallCommand: bun > pnpm > yarn > npm.
-  // package-lock.json is only used to report the lockfile name when no other
-  // lockfile is present; it does not change the install command (npm is the
-  // default fallback regardless).
-  private detectPackageManagerInfo(cwd: string): {
-    packageManager: DevPreviewPackageManager;
-    lockfileName: string | null;
-  } {
-    if (existsSync(path.join(cwd, "bun.lock"))) {
-      return { packageManager: "bun", lockfileName: "bun.lock" };
-    }
-    if (existsSync(path.join(cwd, "bun.lockb"))) {
-      return { packageManager: "bun", lockfileName: "bun.lockb" };
-    }
-    if (existsSync(path.join(cwd, "pnpm-lock.yaml"))) {
-      return { packageManager: "pnpm", lockfileName: "pnpm-lock.yaml" };
-    }
-    if (existsSync(path.join(cwd, "yarn.lock"))) {
-      return { packageManager: "yarn", lockfileName: "yarn.lock" };
-    }
-    if (existsSync(path.join(cwd, "package-lock.json"))) {
-      return { packageManager: "npm", lockfileName: "package-lock.json" };
-    }
-    return { packageManager: "npm", lockfileName: null };
-  }
-
   async getDestructivePreviewMeta(
     request: DevPreviewSessionRequest
   ): Promise<DevPreviewDestructivePreviewMeta> {
     validateSessionRequest(request);
     const cwd = this.resolveSessionCwd(request);
     const [cacheDirs, nodeModules] = await Promise.all([
-      Promise.all(CACHE_DIRS.map((relPath) => this.statDirMeta(cwd, relPath))),
-      this.statDirMeta(cwd, "node_modules"),
+      Promise.all(CACHE_DIRS.map((relPath) => statDirMeta(cwd, relPath))),
+      statDirMeta(cwd, "node_modules"),
     ]);
-    const pmInfo = this.detectPackageManagerInfo(cwd);
+    const pmInfo = detectPackageManagerInfo(cwd);
     return {
       cwd,
       cacheDirs,
@@ -2262,11 +2151,11 @@ export class DevPreviewSessionService {
     const [cacheDirEntries, nodeModulesSizeBytes] = await Promise.all([
       Promise.all(
         CACHE_DIRS.map(async (relPath) => {
-          const size = await this.computeDirSize(path.join(cwd, relPath));
+          const size = await computeDirSize(path.join(cwd, relPath));
           return [relPath, size] as const;
         })
       ),
-      skipNodeModules ? Promise.resolve(null) : this.computeDirSize(path.join(cwd, "node_modules")),
+      skipNodeModules ? Promise.resolve(null) : computeDirSize(path.join(cwd, "node_modules")),
     ]);
     return {
       cacheDirSizes: Object.fromEntries(cacheDirEntries),
@@ -2283,95 +2172,6 @@ export class DevPreviewSessionService {
       );
     }
     return session.cwd;
-  }
-
-  private async statDirMeta(cwd: string, relPath: string): Promise<DevPreviewDirMeta> {
-    const fullPath = path.join(cwd, relPath);
-    try {
-      const stat = await fsPromises.stat(fullPath);
-      if (!stat.isDirectory()) {
-        return { relPath, exists: false, mtimeMs: null };
-      }
-      return { relPath, exists: true, mtimeMs: stat.mtimeMs };
-    } catch {
-      return { relPath, exists: false, mtimeMs: null };
-    }
-  }
-
-  private async computeDirSize(dirPath: string): Promise<number | null> {
-    try {
-      const stat = await fsPromises.stat(dirPath);
-      if (!stat.isDirectory()) return null;
-    } catch {
-      return null;
-    }
-    if (process.platform === "win32") {
-      return this.computeDirSizeWalk(dirPath);
-    }
-    return this.computeDirSizeDu(dirPath);
-  }
-
-  // POSIX path: shell out to `du -sk` (kilobytes). execFile avoids shell
-  // interpretation entirely — dirPath cannot inject. `du` deduplicates
-  // hardlinks by inode within the traversal, so pnpm-linked files inside
-  // node_modules count once per inode (still reports apparent content size,
-  // not the disk that would actually be reclaimed by deletion).
-  private async computeDirSizeDu(dirPath: string): Promise<number | null> {
-    return new Promise((resolve) => {
-      execFile(
-        "du",
-        ["-sk", dirPath],
-        { maxBuffer: 1024 * 1024, timeout: 30_000 },
-        (err, stdout) => {
-          if (err) return resolve(null);
-          const firstField = stdout.trim().split(/\s+/, 1)[0];
-          const kilobytes = Number.parseInt(firstField, 10);
-          if (!Number.isFinite(kilobytes)) return resolve(null);
-          resolve(kilobytes * 1024);
-        }
-      );
-    });
-  }
-
-  // Windows fallback: no `du` available. Walks the tree with opendir,
-  // accumulating file sizes via lstat. Tracks an inode set to dedupe
-  // hardlinks where the FS reports a stable inode (NTFS returns 0/unstable
-  // values in some cases — those entries simply count each time, accepting a
-  // small over-report on edge filesystems).
-  private async computeDirSizeWalk(dirPath: string): Promise<number | null> {
-    try {
-      let total = 0;
-      const seenInos = new Set<string>();
-      const stack: string[] = [dirPath];
-      while (stack.length > 0) {
-        const cur = stack.pop()!;
-        let handle: Awaited<ReturnType<typeof fsPromises.opendir>>;
-        try {
-          handle = await fsPromises.opendir(cur);
-        } catch {
-          continue;
-        }
-        for await (const ent of handle) {
-          const full = path.join(cur, ent.name);
-          if (ent.isDirectory()) {
-            stack.push(full);
-          } else if (ent.isFile() || ent.isSymbolicLink()) {
-            try {
-              const st = await fsPromises.lstat(full);
-              const inoKey = st.ino ? `${st.dev}:${st.ino}` : null;
-              if (inoKey && seenInos.has(inoKey)) continue;
-              if (inoKey) seenInos.add(inoKey);
-              total += st.size;
-            } catch {
-              // skip unreadable
-            }
-          }
-        }
-      }
-      return total;
-    } catch {
-      return null;
-    }
   }
 
   private pollServerReadiness(
